Steps for the PR merge: - Split utils in different files and separate torch specific functions - Refactor a bit - Guard the eval() when there is a risk - Comment many functions
